+++ open-auth/Client/Exception/FactorizeClientProviderNotFoundException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Heptacom\OpenAuth\Client\Exception;
+
+class FactorizeClientProviderNotFoundException extends FactorizeClientException
+{
+}